associate CARD DOM element with index of array
using 'data-' attribute by counting all cases of .card divs and
assigning the data value to card.

as the length of .card divs should always equal index of myLibrary array,
it makes sense to use the two to assign to each other.

next step is to add function to remove array and card when pressing a button.